# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Generation Service",
    "description": "Core services for template processing and LLM interaction.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs/": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create Job",
        "operationId": "createJob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Read Jobs",
        "operationId": "readJobs",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobRead"
                  },
                  "title": "Response Readjobs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Read Job",
        "operationId": "readJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobReadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Jobs"
        ],
        "summary": "Update Job",
        "operationId": "updateJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Delete Job",
        "operationId": "deleteJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/updates": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Add Job Update",
        "description": "Add an update entry to a job's timeline.",
        "operationId": "addJobUpdate",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdateEntry"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClientReadMinimal": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "is_individual": {
            "type": "boolean",
            "title": "Is Individual",
            "default": false
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "ClientReadMinimal"
      },
...
      "FeeType": {
        "type": "string",
        "enum": [
          "fixed",
          "hourly",
          "mixed"
        ],
        "title": "FeeType"
      },
...
      "FileRead": {
        "properties": {
          "file_name": {
            "type": "string",
            "maxLength": 512,
            "title": "File Name"
          },
          "mime_type": {
            "type": "string",
            "maxLength": 255,
            "title": "Mime Type"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "uploaded_by_user_id": {
            "type": "integer",
            "title": "Uploaded By User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "role": {
            "$ref": "#/components/schemas/FileRole"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "instruction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction Id"
          },
          "survey_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Survey Id"
          },
          "preview_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview File Id"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "mime_type",
          "id",
          "storage_key",
          "uploaded_by_user_id",
          "created_at",
          "role"
        ],
        "title": "FileRead"
      },
      "FileRole": {
        "type": "string",
        "enum": [
          "template",
          "preview_pdf",
          "artefact",
          "input"
        ],
        "title": "FileRole"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstructionReadWithInstructionType": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate",
            "default": 0.0
          },
          "forecasted_billable_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecasted Billable Hours",
            "default": 0.0
          },
          "actual_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Hours",
            "default": 0.0
          },
          "contingency_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contingency Percentage",
            "default": 0.0
          },
          "forecasted_settlement_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecasted Settlement Amount",
            "default": 0.0
          },
          "final_settlement_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Settlement Amount",
            "default": 0.0
          },
          "forecasted_fee_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecasted Fee Amount",
            "default": 0.0
          },
          "fee_type": {
            "$ref": "#/components/schemas/FeeType"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstructionStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "updates": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updates"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "instruction_type_id": {
            "type": "integer",
            "title": "Instruction Type Id"
          },
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "lead_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "instruction_type": {
            "$ref": "#/components/schemas/InstructionTypeRead"
          }
        },
        "type": "object",
        "required": [
          "name",
          "fee_type",
          "id",
          "instruction_type_id",
          "job_id",
          "created_at",
          "updated_at",
          "instruction_type"
        ],
        "title": "InstructionReadWithInstructionType"
      },
      "InstructionStatus": {
        "type": "string",
        "enum": [
          "planned",
          "active",
          "completed",
          "archived"
        ],
        "title": "InstructionStatus"
      },
...
      "InstructionTypeRead": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate",
            "default": 0.0
          },
          "default_fee_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeeType"
              },
              {
                "type": "null"
              }
            ],
            "default": "fixed"
          },
          "default_contingency_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Contingency Percentage",
            "default": 0.0
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "default_template_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Template File Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "InstructionTypeRead"
      },
...
      "JobCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "updates": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updates"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "lead_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead User Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "client_id"
        ],
        "title": "JobCreate"
      },
      "JobRead": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "updates": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updates"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client": {
            "$ref": "#/components/schemas/ClientReadMinimal"
          },
          "created_by_user": {
            "$ref": "#/components/schemas/UserRead"
          },
          "lead_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "instructions": {
            "items": {
              "$ref": "#/components/schemas/InstructionReadWithInstructionType"
            },
            "type": "array",
            "title": "Instructions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "client",
          "created_by_user",
          "lead_user",
          "created_at",
          "updated_at"
        ],
        "title": "JobRead"
      },
      "JobReadDetail": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "updates": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updates"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client": {
            "$ref": "#/components/schemas/ClientReadMinimal"
          },
          "created_by_user": {
            "$ref": "#/components/schemas/UserRead"
          },
          "lead_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "instructions": {
            "items": {
              "$ref": "#/components/schemas/InstructionReadWithInstructionType"
            },
            "type": "array",
            "title": "Instructions"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileRead"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "client",
          "created_by_user",
          "lead_user",
          "created_at",
          "updated_at",
          "instructions",
          "files"
        ],
        "title": "JobReadDetail"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "planned",
          "active",
          "completed",
          "archived"
        ],
        "title": "JobStatus"
      },
      "JobUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "lead_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead User Id"
          },
          "updates": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updates"
          }
        },
        "type": "object",
        "title": "JobUpdate"
      },
      "JobUpdateEntry": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "JobUpdateEntry"
      },
...
      "UserRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "org_id": {
            "type": "integer",
            "title": "Org Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "id",
          "org_id",
          "role"
        ],
        "title": "UserRead"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "member",
          "admin"
        ],
        "title": "UserRole"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}